requirements.txt
streamlit>=1.20.0
pillow>=9.0.0
torch>=1.10.0
transformers>=4.25.0 